refactor: Move to entity store
