+++ lib/core.dart
@@ -10,5 +10,5 @@ export 'src/constants.dart';
 export 'src/internal/services/http.dart';
 export 'src/internal/entities/event.dart' show Event, Events, MineralEvent;
 export 'src/internal/entities/command.dart' show Command, MineralCommand, Option, OptionType, OptionChoice, Subcommand, CommandGroup;
-export 'src/internal/managers/store_manager.dart' show Store, MineralStore;
+export 'src/internal/entities/store.dart' show Store, MineralStore;
 export 'src/internal/managers/module_manager.dart' show Module, MineralModule;
